added: search bar
