+++ frontend/src/components/contentBlocks/MyAds/SearchBar/SearchInput/index.js
@@ -1,0 +1,10 @@
+import 'bootstrap/dist/css/bootstrap.css';
+import '../../index.css'
+
+function SearchInput(props) {
+    return <input className="search__input-field mx-1" placeholder={props.placeholder} />
+}
+
+/*"Ads tittle, keyword"*/
+
+export default SearchInput
