+++ frontend/src/components/contentBlocks/MyAds/SearchBar/SearchSelectDropdownMenu/index.js
@@ -1,5 +1,5 @@
 import 'bootstrap/dist/css/bootstrap.css';
-import './index.css'
+import '../../index.css'
 import React, { useRef, useEffect, useState } from "react";
 import SearchSelectDropdownMax from './SearchSelectDropDownMax';
 import SearchSelectDropdownMin from './SearchSelectDropDownMin';
@@ -32,24 +32,26 @@ function SearchSelectDropdownMenu({ props }) {
 
     if (useScreenSize().innerWidth < 640) {
         return (
-            <>
+            <div
+                className="search__dropdown d-flex flex-row flex-nowrap justify-content-between justify-content-md-end align-items-start">
                 {
                     props.map((category, key) => {
-                        return <SearchSelectDropdownMin props={category} key={key}/>
+                        return <SearchSelectDropdownMin props={category} key={key} />
                     })
                 }
-            </>
+            </div>
 
         )
     } else {
         return (
-            <>
-            {
-                props.map((category, key) => {
-                    return <SearchSelectDropdownMax props={category} key={key}/>
-                })
-            }
-        </>
+            <div
+                className="search__dropdown d-flex flex-row flex-nowrap justify-content-between justify-content-md-end align-items-start">
+                {
+                    props.map((category, key) => {
+                        return <SearchSelectDropdownMax props={category} key={key} />
+                    })
+                }
+            </div>
         )
     }
