+++ frontend/src/components/contentBlocks/MyAds/index.css
@@ -1,8 +1,8 @@
 @import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,wght@0,200;0,300;0,400;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,600;1,700;1,800;1,900&family=Nunito:ital,wght@0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
 
 * {
-    margin: 20px;
-    /* margin: 0; */
+    /* margin: 20px; */
+    margin: 0;
     padding: 0;
     font-family: 'Nunito', sans-serif;
     font-style: normal;
@@ -408,7 +408,6 @@ body {
     .search-block {
         /* max-width: 578px;
         font-size: 10px; */
-
         max-width: 100%;
         font-size: 13px;
         gap: 4px;
@@ -599,8 +598,8 @@ body {
         scroll-behavior: smooth;
     }
 
-
     .dropdown-content {
+        top: 125px;
         width: 100%;
     }
     .ads-navigation {
@@ -627,5 +626,9 @@ body {
     .ads__card__buttons {
         justify-content: space-evenly;
         width: 100%;
+    }
+
+    .search-block {
+        gap: 10px;
     }
 }
